Create and manage a Security Group for Trove instances by default
Currently, the charm can be configured with a Neutron Security Group
through the `management-security-groups` config option, which is then
conditionally added to the generated `trove.conf` file.

With this, if the config option is not set, the charm will create and
manage a Security Group containing only egress rules for the related
RabbitMQ units.