Create and manage a Security Group for Trove instances by default
Currently, the charm can be configured with a Neutron Security Group
through the `management-security-groups` config option, which is then
conditionally added to the generated `trove.conf` file.

With this, if the config option is not set, the charm will create and
manage a Security Group containing only egress rules for the related
RabbitMQ units.

Skip having the leader update the security groups if it's departing.
This will speed up the departing process by having less to do, and
it avoid other potential issues like having the keystone relation
already be departed and not have the necessary data to initialize
a client, resulting in an error state.

Sync charm peers by using leadership.leader_set flags.